Made quick_start.sh and setup.sh more user friendly.

+++ setup.sh
@@ -8,15 +8,38 @@ function die() {
 	exit 1
 }
 
-set -x
+# set -x
+
+echo 'This script will either create or update the following files:'
+echo ''
+echo '    ~/.bashrc'
+echo '    ~/.vimrc'
+echo '    ~/.screenrc'
+echo '    ~/.inputrc'
+echo '    ~/.bashrc'
+[ `uname` = 'Darwin' ] && {
+	echo '    ~/Library/KeyBindings/DefaultKeyBinding.dict'
+}
+echo ''
+echo 'If the file already exists, it will be non-destructively'
+echo 'modified by appending commented-out commands. In this case'
+echo 'you will need to edit the respective file manually to'
+echo 'enable shattings.'
+echo ''
+
+read -p 'Press any key to continue setup or CTRL-C to abort.'
 
 if [ -e "$HOMEDIR/.bashrc" ]
 then
-	{
-		echo '# Uncomment to enable'
-		echo '# '". \"$SHATDIR/bash/bashrc\""
-	} >> "${HOMEDIR}/.bashrc"
-	echo Added stub to existing $HOMEDIR/.bashrc
+	if grep -q "$SHATDIR/bash/bashrc" $HOMEDIR/.bashrc
+	then echo $HOMEDIR/.bashrc is already good to go.
+	else
+		{
+			echo '# Uncomment to enable'
+			echo '# '". \"$SHATDIR/bash/bashrc\""
+		} >> "${HOMEDIR}/.bashrc"
+		echo Added stub to existing $HOMEDIR/.bashrc, make sure you edit it
+	fi
 else
 	{
 		echo ". \"$SHATDIR/bash/bashrc\""
@@ -26,11 +49,15 @@ fi
 
 if [ -e "$HOMEDIR/.profile" ]
 then
-	{
-		echo '# Uncomment to enable'
-		echo '# '". ~/.bashrc"
-	} >> "${HOMEDIR}/.profile"
-	echo Added stub to existing $HOMEDIR/.profile
+	if grep -q ". ~/.bashrc" $HOMEDIR/.profile
+	then echo $HOMEDIR/.profile is already good to go.
+	else
+		{
+			echo '# Uncomment to enable'
+			echo '# '". ~/.bashrc"
+		} >> "${HOMEDIR}/.profile"
+		echo Added stub to existing $HOMEDIR/.profile, make sure you edit it
+	fi
 else
 	{
 		echo ". ~/.bashrc"
@@ -40,11 +67,15 @@ fi
 
 if [ -e "${HOMEDIR}/.vimrc" ]
 then
-	{
-		echo '" Uncomment to enable'
-		echo '" '"source $SHATDIR/vim/vimrc"
-	} >> "${HOMEDIR}/.vimrc"
-	echo Added stub to existing $HOMEDIR/.vimrc
+	if grep -q "$SHATDIR/vim/vimrc" $HOMEDIR/.vimrc
+	then echo $HOMEDIR/.vimrc is already good to go.
+	else
+		{
+			echo '" Uncomment to enable'
+			echo '" '"source $SHATDIR/vim/vimrc"
+		} >> "${HOMEDIR}/.vimrc"
+		echo Added stub to existing $HOMEDIR/.vimrc, make sure you edit it
+	fi
 else
 	{
 		echo "source $SHATDIR/vim/vimrc"
@@ -54,7 +85,7 @@ fi
 
 if [ -d "${HOMEDIR}/.vim" ]
 then
-	echo WARNING: "$HOMEDIR/.vim" already exists, skipping...
+	echo  "$HOMEDIR/.vim" already exists, skipping...
 else
 	ln -s "${SHATDIR}/vim" "${HOMEDIR}/.vim"
 	echo Linked $HOMEDIR/.vim
@@ -62,11 +93,15 @@ fi
 
 if [ -e "${HOMEDIR}/.screenrc" ]
 then
-	{
-		echo '# Uncomment to enable'
-		echo '# '"source \"$SHATDIR/screen/screenrc\""
-	} >> "${HOMEDIR}"/.screenrc
-	echo Added stub to existing $HOMEDIR/.screenrc
+	if grep -q "$SHATDIR/screen/screenrc" $HOMEDIR/.screenrc
+	then echo $HOMEDIR/.screenrc is already good to go.
+	else
+		{
+			echo '# Uncomment to enable'
+			echo '# '"source \"$SHATDIR/screen/screenrc\""
+		} >> "${HOMEDIR}"/.screenrc
+		echo Added stub to existing $HOMEDIR/.screenrc, make sure you edit it
+	fi
 else
 	{
 		echo "source \"$SHATDIR/screen/screenrc\""
@@ -76,11 +111,15 @@ fi
 
 if [ -e "${HOMEDIR}/.inputrc" ]
 then
-	{
-		echo '# Uncomment to enable'
-		echo '# '"\$include $SHATDIR/etc/inputrc\""
-	} >> "${HOMEDIR}"/.inputrc
-	echo Added stub to existing $HOMEDIR/.inputrc
+	if grep -q "$SHATDIR/etc/inputrc" $HOMEDIR/.inputrc
+	then echo $HOMEDIR/.inputrc is already good to go.
+	else
+		{
+			echo '# Uncomment to enable'
+			echo '# '"\$include $SHATDIR/etc/inputrc\""
+		} >> "${HOMEDIR}"/.inputrc
+		echo Added stub to existing $HOMEDIR/.inputrc, make sure you edit it
+	fi
 else
 	{
 		echo "\$include $SHATDIR/etc/inputrc"
@@ -88,7 +127,14 @@ else
 	echo Created $HOMEDIR/.inputrc
 fi
 
-[ -e ~/Library/KeyBindings/DefaultKeyBinding.dict ] && echo Skipping ~/Library/KeyBindings/DefaultKeyBinding.dict ||
-ln -s ../../.shattings/Library/KeyBindings/DefaultKeyBinding.dict ~/Library/KeyBindings/DefaultKeyBinding.dict
+[ `uname` = 'Darwin' ] && {
+	if [ -e ~/Library/KeyBindings/DefaultKeyBinding.dict ]
+	then
+		echo Skipping ~/Library/KeyBindings/DefaultKeyBinding.dict
+	else
+		ln -s ../../.shattings/Library/KeyBindings/DefaultKeyBinding.dict ~/Library/KeyBindings/DefaultKeyBinding.dict
+	fi
+}
 
-echo Done!
+echo 'Done! Shattings is now set up.'
+echo 'Remember to edit the files mentioned above, if any.'
